requirements.txt
Django==1.4
SQLAlchemy==0.7.8
djangorestframework==2.3.13
mock==1.0.1
nose==1.3.2
wsgiref==0.1.2
drf-nested-routers==0.1.3
